refactor : 분리된 컨트롤러를 합체
