+++ src/main/java/lotto/Application.java
@@ -1,11 +1,11 @@
 package lotto;
 
-import lotto.controller.Controller;
+import lotto.controller.LottoController;
 
 public class Application {
 
     public static void main(String[] args) {
-        Controller controller = new Controller();
-        controller.play();
+        LottoController lottoController = new LottoController();
+        lottoController.play();
     }
 }
